+++ src/main/java/step2/domain/ResultAnalyzer.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -7,11 +7,12 @@ import java.util.stream.Collectors;
  * 로또 분석 기능을 정의한 클래스
  */
 public class ResultAnalyzer {
-    
+
     private ResultAnalyzer() {}
 
     /**
      * 당첨 된 로또 갯수를 반환한다.
+     *
      * @param winningNumbers
      * @param tickets
      * @param hitCount
@@ -26,6 +27,7 @@ public class ResultAnalyzer {
 
     /**
      * 당첨금을 반환한다.
+     *
      * @param ranking
      * @param winningNumbers
      * @param tickets
@@ -37,6 +39,7 @@ public class ResultAnalyzer {
 
     /**
      * 수익률을 반환한다.
+     *
      * @param winningNumbers
      * @param tickets
      * @param spending
