+++ src/main/java/step2/domain/Ticket.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import java.util.*;
 import java.util.stream.Collectors;
