+++ src/main/java/step2/view/InputScanner.java
@@ -1,4 +1,6 @@
-package step2;
+package step2.view;
+
+import step2.lib.PrintMessage;
 
 import java.util.Arrays;
 import java.util.Scanner;
@@ -10,8 +12,7 @@ public class InputScanner {
 
     private final static Scanner SCANNER = new Scanner(System.in);
 
-    private InputScanner() {
-    }
+    private InputScanner() {}
 
     /**
      * message를 받아 출력과 동시에 사용자 int 형 입력 값을 받는다.
